Day8/pyproject.toml
[project]
name = "day8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipython>=9.6.0",
    "openai>=2.6.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
